Suggest excluding .git files during copy
If the user is copying the files from the template directory to a new
directory they probably don't want to bring along all the .git files as
well, and they might already have .git type files setup in the new
directory that they don't want overwritten. Instead of trying to do
something fancy with `cp` it is easier to just use `rsync`'s
`--exclude` flag.