Use the notoccite package
As the package name alludes to, references in figure or table captions
that thus first appear in the table of contents will not be cited there,
but rather the first time that they appear in the main body of text.

c.f. https://ctan.org/pkg/notoccite?lang=en

Additionally, the graphicpath is set.