Use latexmk for compilation
Borrowing components of the ATLAS LaTeX Makefile, use latekmk to be able
to detect when components of the LaTeX source have chagned and build
then.
c.f. http://mg.readthedocs.io/latexmk.html

latexmk is called explicitly in target `final` to avoid a conflict with
the TERM variable in Circle CI from having clear be called.