Have Makefile compile figures
Have the Makefile find all .tex files under the images directory and
then build them. Their output appears in the top level directory, and is
then moved to the same directory as the source (.tex) file. The
associated auxiliary files are removed.

This allows for .pdf versions of the images to be inserted into the main
text, but only requires that the source .tex file be stored in the repo.