Apply settable to move user data to user_config.tex
Using the settable function all the metadata that is needed to properly
create the approval and title pages of the thesis can be supplied by the
user in a single user controlled file, user_config.tex. user_config.tex
is actually setting the return values of settable functions that have
been defined in DL_thesis.sty or initiated in front_pages.tex.
This effectivley seperates the thesis template into files that define
the style of the docuement, which should be centrally controlled, and
end user editable files.

As part of this, the dedication text now becomes its own end user file.