Add number of previous degrees to user_config
By having the number of previous degrees be controlled from within
user_config.tex this gives the end user easy direct access to
controlling the degrees shown without having to change anything in the
underlying source.

To add another degree they simply declare the degree number and then
fill out the settable properties that are associated with it.