Add arXiv target to Makefile
The arXiv target prepares a tar.gz file with all the necessary files for
arXiv to build the document correctly.

If hyperref is loaded then there is an option clash.
c.f.
https://tex.stackexchange.com/questions/329461/arxiv-option-clash-for-package-hyperref-when-there-are-no-options

arXiv needs a .bbl to generate the bibliography, so compile to generate
it and then move that into the directory to tar as ms.bbl