Remove files from tar.gz that cause conflicts with arXiv
If two files exist that have the same name with the execption of their
extension and they are .pdf and .tex then the arXiv unpacker will delete
the .pdf verion of the file. To avoid this, for image files that are
generated from TikZ remove thier .tex pdf file.

So for something like example.tex and example.pdf remove example.tex

Additionally remove any instances of .bib files. These must be complied
down to .bbl files and arXiv removes any instance of .bib files during
unpacking.