Fix TOC generation and add Glossary to TOC (#45)
* Remove unnecessary numfigure and numtable counts

The unnecessary writing of these to a temporary file (for uses unknown) as part of the \thesis environment cause conflicts with the Table of Contents generation if a Glossary is present

* Add Glossary to Table of Contents

Use the [toc] option for the glossaries package. To get the Glossary to appear capitalized in the TOC use the title and toctile options of \printglossary to set a custom name
c.f. https://en.wikibooks.org/wiki/LaTeX/Glossary#Custom_Name

* Use setglossarystyle command

Avoids warning of:
Package glossaries Warning: Deprecated command \glossarystyle. I recommend you switch to \setglossarystyle unless you want to maintain backward compatibility on input line 36.