Revise abstract page to use minipages
Using \hbox is a bad idea beyond the fact that it is a low level TeX
primitive that shouldn't be used for the most part. It also creates
boxes that don't obey normal formatting rules. So it is much better to
use alignment environments that will be figure out how to do this for
you with minipages to split the left aligned and right aligned segments.