Dynamically determine degree minipage width
Determine the name of the thesis author and then use its width to
dynamically set the width of the minipages for the author name and
degrees on the abstract page. The commands and lengths that are set are
properly scoped so that they are not accessible outside of the
environment.